// src/graphql/generated/types.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Upload: any;
};
// ...
export type Bio = {
  __typename?: "Bio";
  status?: Maybe<Scalars["String"]>;
  time?: Maybe<Scalars["Float"]>;
};

export type BioDto = {
  __typename?: "BioDto";
  status?: Maybe<Scalars["String"]>;
  time?: Maybe<Scalars["Float"]>;
};

export type BioDtoForUpdate = {
  status?: InputMaybe<Scalars["String"]>;
  time?: InputMaybe<Scalars["Float"]>;
};

export type BlockedContact = {
  __typename?: "BlockedContact";
  _id: Scalars["ID"];
  pid: Scalars["String"];
  userId: Scalars["String"];
};

export type BlockedRoom = {
  __typename?: "BlockedRoom";
  blocked_at: Scalars["Float"];
  pid: Scalars["String"];
  reason?: Maybe<Scalars["String"]>;
  room_Id: Scalars["String"];
};
// ...
export type DeviceInfo = {
  __typename?: "DeviceInfo";
  appVersion: Scalars["String"];
  deviceModel: Scalars["String"];
  platform: Scalars["String"];
};
// ...
export enum GlobalFrequencyUnit {
  /** DAY */
  Day = "DAY",
  /** HOUR */
  Hour = "HOUR",
  /** MINUTE */
  Minute = "MINUTE",
  /** WEEK */
  Week = "WEEK",
}

export type GlobalFrequencyUnitInput = {
  Count?: Scalars["Float"];
  Unit?: GlobalFrequencyUnit;
};

export type GlobalFrequencyUnitSchema = {
  __typename?: "GlobalFrequencyUnitSchema";
  Count: Scalars["Float"];
  Unit: GlobalFrequencyUnit;
  _id?: Maybe<Scalars["String"]>;
};
// ...
export type LastSynced = {
  __typename?: "LastSynced";
  contactList: Scalars["DateTime"];
};
// ...
export type ReplyMessage = {
  __typename?: "ReplyMessage";
  cid?: Maybe<Scalars["String"]>;
  created_at?: Maybe<Scalars["Float"]>;
  file_URL?: Maybe<Scalars["String"]>;
  fontStyle?: Maybe<Scalars["String"]>;
  index?: Maybe<Scalars["String"]>;
  message?: Maybe<Scalars["String"]>;
  sender?: Maybe<Scalars["String"]>;
  type?: Maybe<Scalars["String"]>;
};
// ...
export type User = {
  __typename?: "User";
  WebDeviceId?: Maybe<Scalars["String"]>;
  _id: Scalars["ID"];
  appVersion: Scalars["String"];
  bio: Bio;
  blockedRooms: Array<BlockedRoom>;
  contact_reminder?: Maybe<Array<UserContact>>;
  createdAt?: Maybe<Scalars["Float"]>;
  device?: Maybe<DeviceInfo>;
  email: Scalars["String"];
  emailConfirmed: Scalars["Boolean"];
  favouriteChats: Array<FavouriteChat>;
  firstName: Scalars["String"];
  folders: Array<Folder>;
  globalFrequency: GlobalFrequencyUnitSchema;
  isAgree: Scalars["Boolean"];
  isSurvey?: Maybe<Scalars["Boolean"]>;
  iso_code?: Maybe<Scalars["String"]>;
  joinedRooms: Array<JoinedRooms>;
  language?: Maybe<Scalars["String"]>;
  lastDevice?: Maybe<Scalars["String"]>;
  lastName: Scalars["String"];
  lastSeen?: Maybe<Scalars["Float"]>;
  lastSynced: LastSynced;
  loggedIn?: Maybe<Scalars["String"]>;
  loggedOut?: Maybe<Scalars["String"]>;
  mode: UserModeStatus;
  onlineTime?: Maybe<Scalars["String"]>;
  orgId: Scalars["ID"];
  phone: Scalars["String"];
  phoneConfirmed: Scalars["Boolean"];
  profile_img: Scalars["String"];
  receipts: Scalars["Boolean"];
  seniorCitizenRoom: Array<SeniorCitizenRoomSchema>;
  status: Scalars["String"];
  surveyFrequency?: Maybe<Scalars["Float"]>;
  timezone: Scalars["String"];
  updatedAt?: Maybe<Scalars["Float"]>;
  visibility?: Maybe<UserAvailability>;
};

export type UserContact = {
  __typename?: "UserContact";
  CustomMessage: Scalars["String"];
  _id: Scalars["String"];
  firstName: Scalars["String"];
  frequency: OnlineStatusFrequency;
  isDismiss?: Maybe<Scalars["Boolean"]>;
  lastName: Scalars["String"];
  onlineTime?: Maybe<Scalars["String"]>;
  phone: Scalars["String"];
  profile_img: Scalars["String"];
  roomId?: Maybe<Scalars["String"]>;
};
// ...
export type FavouriteChat = {
  __typename?: "favouriteChat";
  cid: Scalars["String"];
  created_at: Scalars["String"];
  fileURL: Scalars["String"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phone: Scalars["String"];
  reply_msg?: Maybe<Array<ReplyMessage>>;
  roomId: Scalars["String"];
  sender: Scalars["String"];
  type: Scalars["String"];
};

export type Folder = {
  __typename?: "folder";
  _id?: Maybe<Scalars["ID"]>;
  name: Scalars["String"];
  roomId: Array<Scalars["String"]>;
};
// ...
export type JoinedRooms = {
  __typename?: "joinedRooms";
  left_at: Scalars["Float"];
  roomId: Scalars["String"];
  type: Scalars["String"];
};
// ...
export enum OnlineStatusFrequency {
  /** ONCE */
  Once = "ONCE",
  /** REPEAT */
  Repeat = "REPEAT",
}
// ...
export type SeniorCitizenRoomSchema = {
  __typename?: "seniorCitizenRoomSchema";
  roomId?: Maybe<Scalars["String"]>;
  userId?: Maybe<Scalars["String"]>;
};
// ...
export enum UserAvailability {
  /** ANONYMOUS */
  Anonymous = "ANONYMOUS",
  /** AVAILABLE */
  Available = "AVAILABLE",
}
// ...
export enum UserModeStatus {
  /** CLASSIC */
  Classic = "CLASSIC",
  /** SENIORCITIZEN */
  Seniorcitizen = "SENIORCITIZEN",
}